==> 03-cleaning.Rmd <==
# Data transformation

For our scoresheet data, when we were done tabulating all the statistics for all the players in Excel, we exported each tab into a CSV file and then read those CSV files into R Studio.  We made one data frame to hold the hitting statistics and one data frame to hold the pitching statistics.  
  
There was some minor data cleanup to do for these data frames.  There were many cells in Excel that were purposely left empty.  For example, if a hitter never had a homerun in APBA baseball, that cell would have been left blank during the tabulation process, but the blank cell really represented a “0”.
When you import into R Studio from a CSV however, these blanks cells turn into “NA”s, so we needed to use R Studio commands to turn these NAs into the 0s they represent.  

```{r}
library(tidyverse) # for working with data frames
```

## Web Scrapped Batting Table

We need to add the following calculated fields into the web scrapped batting data to better measure the performance of the hitters:

- $\text{AVG}=\dfrac{\text{hits(H)}}{\text{at bats(AB)}}$

- $\text{SLG}=\dfrac{\text{first base(1B)}+\text{second base(2B)}\times2+\text{third base(3B)}\times3+\text{home runs(HR)}\times4}{\text{at bats(AB)}}$

- $\text{OBA}=\dfrac{\text{hits(H)}+\text{bases on balls(BB)}+\text{times hit by a pitch(HBP)}}{\text{at bats(AB)}+\text{bases on balls(BB)}+\text{times hit by a pitch(HBP)}}$

- $\text{SAC}=\text{sacrifice hits(SH)}+\text{sacrifice flies(SF)}$

```{r}

## read in raw data scraped from web
df_batting  <- read_csv('data/raw/data_batting_raw.csv')
df_pitching <- read_csv('data/raw/data_pitching_raw.csv')

## batting data clean up
df_batting['AVG'] <- df_batting$H/df_batting$AB
df_batting['SLG'] <- (df_batting$H-df_batting$`2B`-df_batting$`3B`-df_batting$HR+2*df_batting$`2B`+3*df_batting$`3B`+4*df_batting$HR)/df_batting$AB
df_batting['OBA'] <- (df_batting$H+df_batting$BB+df_batting$HBP)/(df_batting$AB+df_batting$BB+df_batting$HBP)
df_batting['SAC'] <- df_batting$SH+df_batting$SF

df_batting <- 
  df_batting %>% 
  select(playerID, name, hits, 
         AVG, SLG, OBA, G, AB, R, H, RBI, `2B`, `3B`, HR, SAC, SB, CS, SO, BB, HBP, TB) %>% 
  arrange(playerID)

# create file
# write_csv(df_batting, 'data/real/RealBatting.csv')
```

The resulting real-life batting table looks like:

```{r, echo=FALSE}
str(df_batting)
```

## Web Scrapped Pitching Table

Then we need to add the following calculated fields into the web scrapped pitching data to better measure the performance of the pitchers:

- $\text{AVG_against}=\dfrac{\text{hits(H)}}{\text{hits(H)}+\text{innings pitched(IP)}\times3}$


```{r}
## pitching data clean up

df_pitching['ERA'] <- df_pitching$earned_run_avg
df_pitching['AVG_against'] <- df_pitching$H/(df_pitching$H+3*df_pitching$IP)

df_pitching <- 
  df_pitching %>% 
  select(playerID, name, throws, 
         ERA, W, L, SV, AVG_against, G, GS, CG, IP, ER, H, SO, BB, HBP, WP, BK) %>% 
  arrange(playerID)

# create file
# write_csv(df_pitching, 'data/real/RealPitching.csv')

```

The resulting real-life pitching table looks like:

```{r, echo=FALSE}
str(df_pitching)
```

## Cleaning the Combined Data

Once real data was scraped into CSV files (again, one file for hitting and one for pitching), these were also put into their own data frames.  We then consolidated the information from the game hitting and real hitting data frames into the same data frame such that every row in the consolidated hitting data frame had all the relevant information for the same player.  For example, the first row of the hitting data frame had both game and real hitting data for “Derek Jeter”.  

We then proceeded to remove all players from the consolidated hitting data frame that had either 0 at-bats (AB) in the game or 0 AB in real life. For these players, we had no good way to analyze or evaluate how well APBA was representing them.  

For players with 0 AB in the game, there was nothing to compare against real life.  In the game, these players were mainly relief pitchers. Relief pitchers do hit, but it’s rare, and such players didn’t pitch in enough games to eventually get the chance to hit.  
 
For players with 0 AB in real life, there was nothing to compare game data to.  In real life, these players were all pitchers in the American League since, in the American League, pitchers never hit and instead have another player designated to hit for them (a so-called “designated hitter”).  

Once these rows were removed from the consolidated hitting data frame, we looked at the remaining rows (each row representing a player) and proceeded to calculate how many of each interesting statistic the player should have had in the board game in order to be completely real-life accurate.  Statistics we thought were interesting were hits (H), total bases (TB), runs (R), runs batted in (RBI), doubles (2B), triples (3B), homeruns (HR), and stolen bases (SB).  

Batting average (AVG) is a derived statistic equal to H / AB while slugging percentage (SLG) is a derived statistic equal to TB / AB.  This means that AVG$\times$AB = H and SLG$\times$AB = TB, so, to get the number of H a player should have had in the game, we multiply his real-life AVG with the number of AB he had in the game.  Similarly, to get the number of TB a player should have had in the game, we multiply his real-life SLG with the number of AB he had in the game. Here’s the lines from our code that do this (where rhdf2 is our data frame):

```{r, echo=TRUE, eval=FALSE}

rhdf2$H_should_have = rhdf2$AVG * rhdf2$GameAB
rhdf2$TB_should_have = rhdf2$SLG * rhdf2$GameAB

```

We can make our own version of AVG and SLG for the other statistics, based on either AB or games played (G) as seemed appropriate (examples: (real-life doubles) / (real-life at-bats) and (real-life stolen bases) / (real-life games)). Once again, here’s the lines from our code:

```{r, echo=TRUE, eval=FALSE}

rhdf2$R_should_have = rhdf2$R / rhdf2$G * rhdf2$GameG
rhdf2$RBI_should_have = rhdf2$RBI / rhdf2$G * rhdf2$GameG
rhdf2$"2B_should_have" = rhdf2$"2B" / rhdf2$AB * rhdf2$GameAB
rhdf2$"3B_should_have" = rhdf2$"3B" / rhdf2$AB * rhdf2$GameAB
rhdf2$HR_should_have = rhdf2$HR / rhdf2$AB * rhdf2$GameAB
rhdf2$SB_should_have = rhdf2$SB / rhdf2$G * rhdf2$GameG

```

Now, for each hitter, we have the expected number of H, TB, R, RBI, 2B, 3B, HR, and SB he should have had in the game and the actual number of H, TB, R, RBI, 2B, 3B, HR, and SB he had in the game.  Therefore, we can find the deviation from reality for each of these statistics by taking the difference between game and expected versions as follows:  

```{r, echo=TRUE, eval=FALSE}

rhdf2$H_deviation = rhdf2$GameH - rhdf2$H_should_have
rhdf2$TB_deviation = rhdf2$GameTB - rhdf2$TB_should_have
rhdf2$R_deviation = rhdf2$GameR - rhdf2$R_should_have
rhdf2$RBI_deviation = rhdf2$GameRBI - rhdf2$RBI_should_have
rhdf2$"2B_deviation" = rhdf2$Game2B - rhdf2$"2B_should_have"
rhdf2$"3B_deviation" = rhdf2$Game3B - rhdf2$"3B_should_have"
rhdf2$HR_deviation = rhdf2$GameHR - rhdf2$HR_should_have
rhdf2$SB_deviation = rhdf2$GameSB - rhdf2$SB_should_have


```


Now for pitchers, the entire process is quite similar.  We consolidated the information from the game pitching and real pitching data frames into the same data frame such that every row in the consolidated data frame had all the relevant information for the same pitcher.  For example, the first row of the pitching data frame had both game and real pitching data for “Brian Bohanon”.   

For pitchers, we look at the number of innings pitched (IP) or games pitched (G) in the game to determine how many of the other statistics pitchers should have.  The standard baseball derived statistic, “earned run average” (ERA), can be used with IP to determine how many earned runs (ER) the player should have.  Once again, we make up our own similarly derived statistics to determine how many of the other pitching statistics pitchers should have. Here’s the lines from our code:  

```{r, echo=TRUE, eval=FALSE}

rpdf$ER_should_have = rpdf$ERA * rpdf$GameIP / 9
rpdf$Hits_should_have = rpdf$H / rpdf$IP * rpdf$GameIP
rpdf$SO_should_have = rpdf$SO / rpdf$IP * rpdf$GameIP
rpdf$W_should_have = rpdf$W / rpdf$G * rpdf$GameG
rpdf$L_should_have = rpdf$L / rpdf$G * rpdf$GameG
rpdf$SV_should_have = rpdf$SV / rpdf$G * rpdf$GameG
rpdf$BB_should_have = rpdf$BB / rpdf$IP * rpdf$GameIP

```

And here’s the lines from our code for calculating the deviation for each pitcher:

```{r, echo=TRUE, eval=FALSE}

rpdf$ER_deviation = rpdf$GameER - rpdf$ER_should_have
rpdf$Hits_deviation = rpdf$GameHits - rpdf$Hits_should_have
rpdf$SO_deviation = rpdf$GameSO - rpdf$SO_should_have
rpdf$W_deviation = rpdf$GameW - rpdf$W_should_have
rpdf$L_deviation = rpdf$GameL - rpdf$L_should_have
rpdf$SV_deviation = rpdf$GameSV - rpdf$SV_should_have
rpdf$BB_deviation = rpdf$GameBB - rpdf$BB_should_have

```


```{r}

#defining dataframes

rhdf <- read_csv('data/real/RealBatting.csv')
rpdf <- read_csv('data/real/RealPitching.csv')

ghdf <- read_csv('data/game/GameBatting.csv')
gpdf <- read_csv('data/game/GamePitching.csv')

#preparing to create new dataframe with both real and game data 
#batting/hitting data

ghdf[is.na(ghdf)] <- 0
rhdf[is.na(rhdf)] <- 0
rhdf$GameAB = ghdf$AB
rhdf$GameH = ghdf$H
rhdf$GameTB = ghdf$TB
rhdf$GameG = ghdf$G
rhdf$GameR = ghdf$R
rhdf$GameRBI = ghdf$RBI
rhdf$Game2B = ghdf$"2B"
rhdf$Game3B = ghdf$"3B"
rhdf$GameHR = ghdf$HR
rhdf$GameSB = ghdf$SB

#only want to work with this condition
rhdf2 <- rhdf %>% filter(GameAB > 0 & AB > 0)

#new columns added

rhdf2$H_should_have = rhdf2$AVG * rhdf2$GameAB
rhdf2$TB_should_have = rhdf2$SLG * rhdf2$GameAB
rhdf2$R_should_have = rhdf2$R / rhdf2$G * rhdf2$GameG
rhdf2$RBI_should_have = rhdf2$RBI / rhdf2$G * rhdf2$GameG
rhdf2$"2B_should_have" = rhdf2$"2B" / rhdf2$AB * rhdf2$GameAB
rhdf2$"3B_should_have" = rhdf2$"3B" / rhdf2$AB * rhdf2$GameAB
rhdf2$HR_should_have = rhdf2$HR / rhdf2$AB * rhdf2$GameAB
rhdf2$SB_should_have = rhdf2$SB / rhdf2$G * rhdf2$GameG

rhdf2$H_deviation = rhdf2$GameH - rhdf2$H_should_have
rhdf2$TB_deviation = rhdf2$GameTB - rhdf2$TB_should_have
rhdf2$R_deviation = rhdf2$GameR - rhdf2$R_should_have
rhdf2$RBI_deviation = rhdf2$GameRBI - rhdf2$RBI_should_have
rhdf2$"2B_deviation" = rhdf2$Game2B - rhdf2$"2B_should_have"
rhdf2$"3B_deviation" = rhdf2$Game3B - rhdf2$"3B_should_have"
rhdf2$HR_deviation = rhdf2$GameHR - rhdf2$HR_should_have
rhdf2$SB_deviation = rhdf2$GameSB - rhdf2$SB_should_have

#added new combined dataframe to files

#write_csv(rhdf2, 'data/combined/Batting.csv')

#pitching data

gpdf[is.na(gpdf)] <- 0
rpdf[is.na(rpdf)] <- 0
rpdf$GameG = gpdf$G
rpdf$GameIP = gpdf$IP
rpdf$GameER = gpdf$ER
rpdf$GameHits = gpdf$Hits
rpdf$GameSO = gpdf$SO
rpdf$GameW = gpdf$Wins
rpdf$GameL = gpdf$Losses
rpdf$GameSV = gpdf$Saves
rpdf$GameBB = gpdf$BB

#new columns added

rpdf$ER_should_have = rpdf$ERA * rpdf$GameIP / 9
rpdf$Hits_should_have = rpdf$H / rpdf$IP * rpdf$GameIP
rpdf$SO_should_have = rpdf$SO / rpdf$IP * rpdf$GameIP
rpdf$W_should_have = rpdf$W / rpdf$G * rpdf$GameG
rpdf$L_should_have = rpdf$L / rpdf$G * rpdf$GameG
rpdf$SV_should_have = rpdf$SV / rpdf$G * rpdf$GameG
rpdf$BB_should_have = rpdf$BB / rpdf$IP * rpdf$GameIP

rpdf$ER_deviation = rpdf$GameER - rpdf$ER_should_have
rpdf$Hits_deviation = rpdf$GameHits - rpdf$Hits_should_have
rpdf$SO_deviation = rpdf$GameSO - rpdf$SO_should_have
rpdf$W_deviation = rpdf$GameW - rpdf$W_should_have
rpdf$L_deviation = rpdf$GameL - rpdf$L_should_have
rpdf$SV_deviation = rpdf$GameSV - rpdf$SV_should_have
rpdf$BB_deviation = rpdf$GameBB - rpdf$BB_should_have

#added new combined dataframe to files

#write_csv(rpdf, 'data/combined/Pitching.csv')

```
