==> 04-missing.Rmd <==
# Missing values

```{r}
#load necessary packages
library(ggplot2)
library(visdat)
library(naniar)
```
Missing values can complicate data exploration, therefore our aim is to find and remove them. After replacing the appropriate NA's with 0, we use an R command that shows how many NA values are in each column of both the batting and pitching data.

```{r}

#working with both combined data sets
#read both data files

combh <- read_csv("data/combined/Batting.csv")
combp <- read_csv("data/combined/Pitching.csv")

```

```{r echo=FALSE}

#calculation of number of NA's in data

colSums(is.na(combh)) #hitting\batting data
colSums(is.na(combp)) #pitching data

```

As expected, we see that there are no NA values in either of the data sets, hence there is no missing data.
Plotting the two data sets asserts this.

```{r}
vis_miss(combh) + 
  labs(title = "Missingness of Batting Data")
vis_miss(combp)+ 
  labs(title = "Missingness of Pitching Data")

gg_miss_var(combh)+ 
  labs(title = "Missingness of Batting Data")
gg_miss_var(combp)+ 
  labs(title = "Missingness of Pitching Data")
```
