==> 06-interactive.Rmd <==
# Interactive component

```{r}
#create new data frames for both batting and pitching

bat <- read.csv('data/combined/Batting.csv')
pit <- read.csv('data/combined/Pitching.csv')

#want to work with names, left or right handed and deviation of hits

bbat <- bat[,c(2,3,40)]
bbat$H_deviation <- signif(bbat$H_deviation,2)

#want to work with names, left or right handed and deviation of ER

bpit <- pit[,c(2,3,36)]
bpit$ER_deviation <- signif(bpit$ER_deviation,2)

#create two new files

#write.csv(bbat, 'data/interactive/Boxbatting.csv')
#write.csv(bpit, 'data/interactive/Boxpitching.csv')

```

## ER Deviation by Left and Right-handed Pitchers

<meta charset="utf-8">

<!-- Load d3.js -->
<script src="https://d3js.org/d3.v4.js"></script>

<!-- Create a div where the graph will take place -->
<div id="my_dataviz"></div>

<!-- Plugin for color scale -->
<script src="https://d3js.org/d3-scale-chromatic.v1.min.js"></script>

<!-- Tooltip style -->
<style>
.tooltip {
  background-color: black;
  border: none;
  border-radius: 5px;
  padding: 15px;
  min-width: 400px;
  text-align: left;
  color: white;
}
</style>

<script>

// set the dimensions and margins of the graph
var margin = {top: 10, right: 30, bottom: 50, left: 70},
    width = 660 - margin.left - margin.right,
    height = 600 - margin.top - margin.bottom;

// append the svg object to the body of the page
var svg = d3.select("#my_dataviz")
  .append("svg")
    .attr("width", width + margin.left + margin.right)
    .attr("height", height + margin.top + margin.bottom)
  .append("g")
    .attr("transform",
          "translate(" + margin.left + "," + margin.top + ")");

// Read the data and compute summary statistics for each specie
d3.csv("https://raw.githubusercontent.com/zichenghuang/baseball-board-game/main/data/interactive/Boxpitching.csv", function(data) {

  // Compute quartiles, median, inter quantile range min and max --> these info are then used to draw the box.
  var sumstat = d3.nest() // nest function allows to group the calculation per level of a factor
    .key(function(d) { return d.throws;})
    .rollup(function(d) {
      q1 = d3.quantile(d.map(function(g) { return g.ER_deviation;}).sort(d3.ascending),.25)
      median = d3.quantile(d.map(function(g) { return g.ER_deviation;}).sort(d3.ascending),.5)
      q3 = d3.quantile(d.map(function(g) { return g.ER_deviation;}).sort(d3.ascending),.75)
      interQuantileRange = q3 - q1
      min = q1 - 1.5 * interQuantileRange
      max = q3 + 1.5 * interQuantileRange
      return({q1: q1, median: median, q3: q3, interQuantileRange: interQuantileRange, min: min, max: max})
    })
    .entries(data)

  // Show the Y scale
  var y = d3.scaleBand()
    .range([ height, 0 ])
    .domain(["Left", "Right"])
    .padding(.4);
  svg.append("g")
    .call(d3.axisLeft(y).tickSize(0))
    .select(".domain").remove()

  // Show the X scale
  var x = d3.scaleLinear()
    .domain([-20,20])
    .range([0, width])
  svg.append("g")
    .attr("transform", "translate(0," + height + ")")
    .call(d3.axisBottom(x).ticks(5))
    .select(".domain").remove()

  // Color scale
  var myColor = d3.scaleSequential()
    .interpolator(d3.interpolateInferno)
    .domain([-6,14])

  // Add X axis label:
  svg.append("text")
      .attr("text-anchor", "middle")
      .attr("x", width/2)
      .attr("y", height + margin.top + 30)
      .text("ER Deviation");

  // Show the main vertical line
  svg
    .selectAll("vertLines")
    .data(sumstat)
    .enter()
    .append("line")
      .attr("x1", function(d){return(x(d.value.min))})
      .attr("x2", function(d){return(x(d.value.max))})
      .attr("y1", function(d){return(y(d.key) + y.bandwidth()/2)})
      .attr("y2", function(d){return(y(d.key) + y.bandwidth()/2)})
      .attr("stroke", "black")
      .style("width", 40)

  // rectangle for the main box
  svg
    .selectAll("boxes")
    .data(sumstat)
    .enter()
    .append("rect")
        .attr("x", function(d){return(x(d.value.q1))}) // console.log(x(d.value.q1)) ;
        .attr("width", function(d){ ; return(x(d.value.q3)-x(d.value.q1))}) //console.log(x(d.value.q3)-x(d.value.q1))
        .attr("y", function(d) { return y(d.key); })
        .attr("height", y.bandwidth() )
        .attr("stroke", "black")
        .style("fill", "#69b3a2")
        .style("opacity", 0.3)

  // Show the median
  svg
    .selectAll("medianLines")
    .data(sumstat)
    .enter()
    .append("line")
      .attr("y1", function(d){return(y(d.key))})
      .attr("y2", function(d){return(y(d.key) + y.bandwidth()/2)})
      .attr("x1", function(d){return(x(d.value.median))})
      .attr("x2", function(d){return(x(d.value.median))})
      .attr("stroke", "black")
      .style("width", 80)

  // create a tooltip
  var tooltip = d3.select("#my_dataviz")
    .append("div")
      .style("opacity", 0)
      .attr("class", "tooltip")
      .style("font-size", "16px")
  // Three function that change the tooltip when user hover / move / leave a cell
  var mouseover = function(d) {
    tooltip
      .transition()
      .duration(200)
      .style("opacity", 1)
    tooltip
        .html("<span style='color:pink'>ER Deviation: </span>" + d.ER_deviation + "<br>" + d.name) // + d.Prior_disorder + "<br>" + "HR: " +  d.HR)
        .style("left", (d3.mouse(this)[0]+30) + "px")
        .style("top", (d3.mouse(this)[1]+30) + "px")
  }
  var mousemove = function(d) {
    tooltip
      .style("left", (d3.mouse(this)[0]+30) + "px")
      .style("top", (d3.mouse(this)[1]+30) + "px")
  }
  var mouseleave = function(d) {
    tooltip
      .transition()
      .duration(200)
      .style("opacity", 0)
  }

  // Add individual points with jitter
  var jitterWidth = 50
  svg
    .selectAll("indPoints")
    .data(data)
    .enter()
    .append("circle")
      .attr("cx", function(d){ return(x(d.ER_deviation))})
      .attr("cy", function(d){ return( y(d.throws) + (y.bandwidth()/2) - jitterWidth/2 + Math.random()*jitterWidth )})
      .attr("r", 4)
      .style("fill", function(d){ return(myColor(+d.ER_deviation)) })
      .attr("stroke", "black")
      .on("mouseover", mouseover)
      .on("mousemove", mousemove)
      .on("mouseleave", mouseleave)


})

</script>

For our interactive component, we wanted to see if APBA did a better job at modeling left-handed pitchers or right-handed pitchers.  We chose earned runs given up (ER) as an example statistic and again looked at the ER deviation from what was expected.  Also, we again chose to use box and whisker plots.

We found that APBA did quite a bit better at modeling right-handed pitchers.  This is evidenced by a smaller-width box which is much more centered on 0.  It appears 50% of right-handed pitchers gave up within just a few ER of what was expected.  Meanwhile, the box representing 50% of left-handed pitchers was not well centered on 0.  It's possible our sample set had lots of players that hit left-handed pitchers well, but from the data we have, it appears APBA biases left-handed pitchers such that they often give up more runs than expected.  Furthermore, not only is box not well centered, but it is also wide.  This means the box representing 50% of left-handed pitchers includes pitchers that gave up as many as 10 runs more than expected!

Note that we chose to make this interactive because we wanted to see immediately which pitchers over-performed or under-performed compared to what was expected.  Hovering over points in the plots allows you to see exactly that.  When you hover over a point and look below the plot (below the bottom axis label, you may have to zoom out), you see the name of the pitcher the point represents and the ER deviation of his performance compared to what is expected (positive or negative).

We considered showing the pitcher name and ER deviation in the middle of the plot when you hover, but we didn’t want to cover other points when showing the information.


This plot was created with the aid of ([ref](https://www.d3-graph-gallery.com/graph/boxplot_horizontal.html)).
